Implement Stray Pokemon Mystery Encounter Added a new mystery encounter in src/data/mystery-encounters/encounters.
This new encounter revolves around an encounter with a helpless
charmander and a Seismitoad.

In this encounter, the mystery encounter utilities are used to define
4 options to the player:

Battle the Seismitoad: Player starts a Boss Battle, sets rewards with
SetEncounterRewards and starts with initBattleWithEnemyConfig, using
the previously defined boss Pokemon.

Protect Charmander: Uses withPrimaryPokemonRequirement to define
the need for PROTECTING_MOVES for this option, in which the player
protects the Charmander and fends off the Seismitoad using
leaveEncounterWithoutBattle to skip the encounter's boss battle.

Distract the Seismitoad: Uses withPrimaryPokemonRequirement to define
the need for DISTRACTION_MOVES for this option, in which the player
distracts the Seismitoad, affecting its stats.

Have your fastest Pokemon take the Charmander: using the
getFastestPokemon function and the defined function to remove
pokemon from party, the game removes the players fastest pokemon,
the Seismitoad's stats are increased by 1 and uses the new function
giveLeadPokemonAttackTypeBoostItem to give the lead pokemon charcoal
if the player wins the battle.

In all of these options, if the player is sucessfull, the function
offerCharmanderToJoin adds the charmander to the players party.

Adds new test stray-pokemon-encounter.test.ts to verify proper
behavior for the added mystery encounter.

Signed-off-by: Matheus Alves <matheus.r.noya.alves@tecnico.ulisboa.pt>
Co-authored-by: Inês Simões <ines.p.simoes@tecnico.ulisboa.pt>